version-a-durable-functions/requirements.txt
azure-functions
azure-durable-functions